pagination

+++ extern-api-java-sdk/src/main/java/ru/kontur/extern_api/sdk/service/DocflowService.java
@@ -550,7 +550,7 @@ public interface DocflowService {
      */
     @Deprecated
     CompletableFuture<QueryContext<DocflowPage>> getDocflowsAsync(Boolean finished,
-            Boolean incoming, long skip, int take, String innKpp, Date updatedFrom, Date updatedTo,
+            Boolean incoming, int skip, int take, String innKpp, Date updatedFrom, Date updatedTo,
             Date createdFrom, Date createdTo, String type);
 
     /**
